Add support for nodejs testing on Debian
This updates the nodejs-test-dependencies role to install packages on
Debian as well as Ubuntu. The main driver for this is Ubuntu Jammy has
updated firefox to run out of a snap and getting selenium to operate in
that environment has been difficult. It is theoretically possible, but
rather than force users to sort it out offer Debian as an alternative.

Change-Id: I51cc95046520597a02c307c5d368f492933ed263