Avoid tox-bindep in test-tox job
The test-tox job exercises the "requires" setting in tox.ini, and
to do so it uses a small package, tox-bindep.  This was originally
added in I521580da978903cd116c872447b1c028c470e63d.

This package is not compatible with tox v4, so to retain this test
we update the test tox.ini file to use another small package,
requestsexceptions, which does not interact with tox and so does
not suffer from this problem.

Change-Id: I6c4737c138bf138afbb4b15d64ac11437c725d4a