Stop trying to use a %s tox target
Tox appears to be very good at always installing the latest version of
itself when you use environment requirements. This was fine until latest
tox stopped accepting the %s environment (but only when you don't
specify the environment explicitly). This %s environment was added to
check that tox siblings behavior didn't interpolate strings, but we use
RawConfigParser now to prevent that and that shouldn't change any time
soon. Testing this otherwise seems painful.

Instead of simply deleting this test we modify it to use a new
acceptable name in order to test the running of a target with a non
default config file path.

Change-Id: Ie92f2a2536ea03ac2058659a083fb23151303ed0