linters-requirements : update Ansible to 2.12, ansible-lint <6.12.0
The recent release of ansible-compat, which is pulled in by
ansible-lint, has dropped support for Ansible < 2.12.  This was pinned
by I233fae8c9036d295968a97ee80e07fde8846c633 and it's not really clear
to me why the upper-bound was added then.

I think we can just uncap this?  It doesn't match Zuul at the moment
anyway which is Ansible 6 (2.13).

Also, since the first revision of this change, ansible-lint 6.12.0
came out.  We are not ready for it, so we need to squash a cap for it
into this change to unbreak the gate.

[1] https://github.com/ansible/ansible-compat/releases/tag/v3.0.0

Change-Id: I901769a133ab3cf68fe314940cd359e54dfd5353