container-roles-jobs: Update tests to jammy nodes
These build tests are all running on Bionic for no particular reason
other than it was recent at the time; update them to Jammy.

Because Jammy appears to have a too old systemd to properly support
podman and non root continers [0] our image builds don't work by
default. We need to override the cgroup manager to cgroupfs which is
less than ideal but should be functional.

[0] https://github.com/containers/podman/issues/7650

Change-Id: I18c56aea776e9ef102fe9020db592fb04df372e2