Fix ensure-docker for docker-ce 23.0.0 upstream packaging
Upstream docker-ce recently released version 23.0.0. This version
appears to depend on apparmor but does not explicitly pull it in.
Not having apparmor results in these errors:

  error: exec: "apparmor_parser": executable file not found in $PATH.

Fix this by adding the apparmor package to the list of packages we
install from the distro for docker. Opensuse also uses apparmor but we
don't support installing docker from upstream on opensuse so we don't
need to make changes to opensuse just Debuntu.

Additionally, buildx appears to have been split out into its own package
now and we need to install it explicitly in order for multiarch CI
builds to be successful.

Change-Id: I7b6fd895f58de9e052af8efca27b9ed4bfac7036