Cap ansible-lint <6.13
Recent ansible-lint releases have started complaining about
"name[play]: All plays should be named." throughout the playbooks in
this repository regardless of whether the commit changes them. This
may be a new difference in how ansible-lint decides which files
should be checked, but it's unclear from the release notes whether
this was intentional or a regression.

Change-Id: I906e889ad2ee50bcdacf7d5e8e0c0895df00e562