promote-image-container: do not delete tags
As noted inline, currently promote-image-container uses skopeo to
delete tags which is not the semantics we want.  This results in the
whole image being removed.

For safety we remove this call; we have two solutions in follow-ons
(deleting tags directly from the registry with a generic tag, and
promoting images from the intermediate registry).

Change-Id: I4b257f593275413da9a50a0cc64e13638e7f94cb