Force cgroupfs cgroup manager with podman on ubuntu
Per https://github.com/containers/podman-compose/issues/209 it seems
this is a known thing with non root containers, podman, and Ubuntu. Just
apply the necessary woraround in a config file so that we don't have to
be aware of it every time we try to use podman in this way. It also
makes using podman-compose simpler.

Change-Id: Ic024e242d99b7f0d4fbd4cf2c3489022b364672d