Use zuul_workspace_root for prepare/mirror workspace test roles
Use zuul_workspace_root as the destination folder for preparing and
mirroring the workspace, which points to ansible_user_dir by default.
(As is currently the case).
The override of this variable can be useful when using a static node
with multiple slots to use a separate workspace for each slot.

Change-Id: I2014b9c876907cfc17a0ec7fa2b6d4e67cb2331c