Drop Helm v2 support to fix v3 issue
Helm v2 is long gone and there seems to be no interest in
supporting it any longer. [1]

Dropping support for v2 allows for an easy fix to the
command used so that it names the release properly
as it was supposed to instead of trying to set the
namespace name (the -n switch changed purpose [1]).

[1] https://lists.zuul-ci.org/archives/list/zuul-discuss@lists.zuul-ci.org/thread/WUWBM5F3PXXDLKK6JNSP4UR4VTWDNPZ4/

Story: 2010783
Task: 48201
Change-Id: I00676bcc7e8b4f3ec014f2292c6c0ee4c17acd96