Fix nox job python2.7 support
Virtualenv dropped python2.7 support. When we install latest nox it also
installs latest virtualenv. Then when we try to run python2.7 jobs which
will create a python2.7 venv things fail with:

  RuntimeError: failed to find interpreter for Builtin discover of python_spec='python2.7'

The underlying issue is that we are getting a too new virtualenv. Fix
this by adding a pre run step to the python2.7 jobs that downgrades to a
version of virtualenv that should support python2.7 environments.

Change-Id: I0ed9915b25c752cdaae4df19c9a0e0dc1aafd3d7