prepare-workspace-git: Add allow deleting current branch
In cases where the HEAD branch of a repository is deleted,
prepare-workspace-git will be unable to propagate that deletion from
the executor repos to a clone of a cached repo on a remote node.
Correct this by unsetting the receive.delyDeleteCurrent flag.  This was
correctly set in mirror-workspace-git, but did not make it into
prepare-workspace-git.

Change-Id: I159fb1f9f9cee873466f11be2f155bed6892472e