Switch fips fole testing to CentOS 9 Stream
CentOS 8 Stream is EOL and these jobs are no longer functional. We would
like to clean up the test nodes entirely as a result of this. Part of
this cleanup is either deleting existing jobs that rely on CentOS 8
Stream or converting them to run on a different platform. In this case
we update the fips role job to run on CentOS 9 Stream.

Change-Id: Ifb7d6deaf529e21c257590d63938026584bee7ae