Make prepare-workspace-git behavior more consistent
This runs the same commands on the git repo regardless of whether
it was cloned or if the role found it there already.  Since the
purpose of the role is to mirror the workspace repos from the
executor to the remote node, this will produce more consistent
behavior.

Note that anyone somehow relying on, say, the origin being set
outside of this role may encounter a behavior change.  It is expected
that anyone manipulating a repo that is also managed by this role
would perform those manipulations idempotently after running this
role.

Change-Id: I428bf2980a526919d5b154c585943be92d4c1cfa