Ignore some recent ansible-list warnings
We don't need to name every play; in the Zuul context, they are
usually pretty self-evident.  It might be nice, but it doesn't
seem necessary to require it.

We really don't need to care about using upper or lower case.

The key-order seems very arbitrary and counterproductive to make
all our developers memorize someone else's arbitrary preference
for ordering.

Change-Id: I49455b6946d5d9b6bffd58420fea586ecc6c5f80