Install doc bindep profile in zuul-jobs-test-tox
This job that tests the zuul-jobs tox role runs various tox targets
including the docs target. This means we need to install dependencies
for doc building. On Ubuntu Noble this include libjpeg-dev for Pillow
wheel building and we fail without this profile installed.

We keep the default profiles of compile and test because we also run
other targets that likely depend on these profiles.

Change-Id: Ifa3495488f35b1fbe4fc665c4d0ac5ed8adb33aa