Fix dockerhub check in use-docker-mirror role
We were checking if dockerhub is a valid key in the
zuul_site_mirror_info/mirror_info dictionaries but did so without
quoting dockerhub as a string. This meant ansible tried to look up
dockerhub as a variable producing this error:

  The conditional check 'dockerhub is in zuul_site_mirror_info' failed.
  The error was: error while evaluating conditional (dockerhub is in
  zuul_site_mirror_info): 'dockerhub' is undefined. 'dockerhub' is undefined

Fix this by quoting dockerhub so that we lookup the string as a key
instead of a variable.

Change-Id: Ie869b9b52fd0a5b70fc07548ce449937ed2c9589