Add ensure-uv role
Uv (https://docs.astral.sh/uv/) is not declared as a dependency for a
Python project, it must be available somehow in the system. This role
installs it if missing.

- Latest version is installed, unless `ensure_uv_version` is
  informed.

- The installed executable path is set as the `uv_executable` fact.

- The `/usr/local/bin/uv` symlink can also be created if
  `ensure_uv_global_symlink: true`.

This new role is a verbatim copy of the `ensure-poetry` role, just doing
a `s/poetry/uv/g`. Even this commit is a replay of the commit adding
that role: 524b7e7b95dcd6adc311e74dd7f0e6da8a3cce58.

Change-Id: I55bc5e1d273045d0978b09f719bf79a875336e30