Replace debian:testing with quay.io/opendevmirror/httpd:alpine
We have several tests that were using debian:testing from docker hub as
a base image to deploy some content that would validate execution of
container runtimes. Docker hub has some pretty strict rate limits in
place these days so we'd like to use an image on quay.io instead.
The opendevmirror org is already mirroring the httpd:alpine image there
which is a smal relatively simple image that we can use for this
purpose.

Note we switch from debian with bash to alpine with a busybox sh. But
the tools we rely on (touch, echo, sleep) all appear to be present.

Change-Id: I9bb5db416e3b9601c67de1c053162fd30a977bbd