Use mirrored qemu-user-static image
Along with the buildkit buildx image we rely on the
multiarch/qemu-user-static image to execute code on an emulated
target architecture for image builds. This image is hosted on docker hub
which has strict rate limits in place so we've mirrored it to
quay.io/opendevmirror. Consume the image from there to reduce impact of
rate limits.

Change-Id: Ie646f0e50cdc23da6ca82568e48f1fa416c35e1d