Update zookeeper latest version to 3.9.3
The ensure-zookeeper role is by default set to install the "latest"
version of zookeeper. Unfortunately, we had that value set to 3.6.1 but
the latest version is actually 3.9.3. Update the latest version to 3.9.3
so that modern Java versions on systems like Noble can run a zookeeper
that is more concurrent with the version of Java available.

I suspect this may help with Zuul's unittest reliability on python3.12
test jobs which run on Noble. I noticed we are installed Java 21 and
Zookeeper 3.6.1 there and then often have flaky connectivity to
Zookeeper. Before digging into Zookeeper reliability it seems important
to be running current versions on top of the current Java 21
installation.

Finally anyone that doesn't want the latest Zookeeper version should be
overriding the version as default is "latest". For this reason I don't
think we need to send notifications or announce this change as long as
it generally works.

Change-Id: Id8491939b174fb13c18da27f027b120704f69c65