Pin artifactory Docker image to 7.77.12
The last successful run for the job `zuul-jobs-test-upload-artifactory`
was almost one year ago in 2024-05-12.

Currently, the artifactory container is unable to start successfully.
The ping task "Wait for artifactory to start" fails with:

    Status code was -1 and not [200]:
    Connection failure: [Errno 104] Connection reset by peer

In this change we're pinning the artifactory image to a version that was
likely to be set as `latest` back in 2024-05-12, according to
https://jfrog.com/help/r/jfrog-release-information/artifactory-7.77.12-self-hosted

This version worked locally, but it took a long time to startup:

    $ docker run -p 8081:8081 -p 8082:8082 --rm \
      docker.bintray.io/jfrog/artifactory-oss:7.77.12
    ...
    All services started successfully in 248.494 seconds

That's why this change is also increasing the delay in the retries.

While trying other more recent versions (7.77.17, 7.84.10), they failed
with:

    Failed pinging artifactory for 300connect ECONNREFUSED 127.0.0.1:8046

Change-Id: I6be6c28f902a73d5f1ff22d47b2d86b8ed9cbf43