Update upload-image-swift
This updates upload-image-swift with some changes that have merged
to the originating role in opendev/zuul-providers.  Hopefully with
this change, we can move opendev to use this role.

This anticipates https://review.opendev.org/956219 by adding a default
value of zuul.image_build_name.  That change is not set as a dependency
because it's not strictly necessary and it's harmless if it merges out
of order (or not at all).

Change-Id: Ib26ae8e54174650a2d082ef2504ef70f4804dbc4