Vendor openvswitch_bridge
The upstream openvswitch collection has been deprecated and was not
included in newer Ansible (like Ansible 11). Vendor the
openvswitch_bridge module from that collection as we use it for
multinode bridge setups. The code was fetched from:

  https://raw.githubusercontent.com/ansible-collections/openvswitch.openvswitch/d375078cfd942599c42fe28e46f59f063c9d3a9d/plugins/modules/openvswitch_bridge.py

And is currently unmodified except for removal of the shebang (to make
linter rules happy), flake8: noqa at the file level (to disable python
linter checks), and a new comment block explaining why the code
was vendored and where it originated from.

A warning is added to the role README indicating that this role is now
effectively deprecated due to its dependencies. An alternative approach
that should be easier to maintain long term is suggested as well.

Change-Id: I2c90d3145b50498b4759046d43b02f70c10715e7