Default configure_mirrors_extra_repos to False in configure-mirrors
Currently configure_mirrors_extra_repos defaults to True. This means
that when configure-mirrors runs we automatically enable extra
repositories on distros like CentOS and Debian. This is particularly
problematic for Debian because we're enabling backports by default, but
Debian will delete the entire backports repo as a release gets older.
This is the case for Debian Bullseye. This means that the default
configuration of this role does not work with Debian Bullseye.

By flipping the default to False we ensure a more upstream like
environment that should generally work over time. Then zuul userse can
opt into enabling these extra repos on a case by case basis if their
jobs require packages from these extra repos. Overall this should be a
safer default for us to avoid mysteriously breaking things in the future
when cleanups happen.

Note this should only be merged after a sufficient amount of time has
passed after we announce this change is going to happen.

Change-Id: I0c60c6004cf1d74a31aff83541d0e3bfcf324e5b