Always build docker images with custom buildx builder
Since docker version 23 buildx is the default builder. We had assumed
that configuring /etc/buildkitd/buildkitd.toml would configure this
default builder. When the buildset registry is in use we added it as a
mirror for common registries under the theory that this would enable
image builds using speculative image builds housed within the buildset
registry. Unfortunately, this doesn't seem to be the case and we have to
configure a custom buildx builder and supply the config to that
builder to get the behavior we want.

The good news is that this is already how multiarch image builds work.
Update the two build-(container|docker)-image roles to always use the
buildx path for docker based image builds. This ensures we configure a
custom builder with a working config, ssl certificates, and /etc/hosts
values enabling the build process to lookup images within buildset
registries. This does add some complication to the build process but it
enables some fundamental zuul functionality so seems worthwhile.

In the process I made a slight simplification refactor to the
build-docker-image role to make its buildx.yaml task list follow the
tasks from build-container-image. This isn't strictly necessary but
simplifies the multiarch vs single arch build paths.

The parent change updated testing to cover this case then marked the
scenarios that did not work (single arch docker build + image hosted
outside of docker.io) as nonvoting. We update those cases to be voting
now that we've fixed these scenarios.

Change-Id: I81c93dd22331b9793f87221eb5b46c9140042be5